agents/pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asyncio>=3.4.3",
    "google-adk>=1.0.0",
    "mcp>=1.9.1",
    "python-dotenv>=1.1.0",
]

[project.scripts]
# Add any scripts here if needed

[tool.uv.sources]
custom_utils = { path = "./custom_utils", editable = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# Define the packages that are part of the 'agents' project.
# This tells Hatch what to include in the wheel.
packages = ["rcsb_agent", "test_agent", "alphafold_agent", "admet_agent"]
